add instructions and code for pre-push githook
This is useful if contributors want to stop failing the CI build link
checker verification.  To validate all links in the docs automatically,
we're using a CI script that runs a few commands on all pull requests.

If this fails on GitHub, the contributor needs to (a) first know/notice
that it failed then (b) make time to fix the broken links and (c) push
again.

To avoid this scheduling and round trip pain, this approach stops a push
if the link checker fails locally first.