remove or clarify all mention of TypeScript
Some unproductive was reported by members of the team around the
assumption that AssemblyScript was as feature rich as TypeScript.

This change aims to remove or explicitly qualify all mention on TypeScript
throughout our documentation.

Where possible, TypeScript was renamed to AssemblyScript. When this was
not possible, some qualification of the relationship between
AssemblyScript and TypeScript was made:

ie. AssemblyScript (a strict subset of TypeScript) ...

Hopefully this eliminates or at least significantly reduces the chances
for confusion between these two technologies.