overhauled initial developer experience in docs (#216)
based on feedback from @SkidanovAlex:

The experience of QuickStart is not very smooth today. As a new person,
I open the docs, click on "Applications Developers", and from there it's
very hard to find an easy-to-follow tutorial. The best I found is
zero-to-hero, but on the left it is called "Building an oracle",
certainly not something I'd click if I was looking for an easy tutorial.

The tutorial on the wallet integration is not a tutorial, it's more of
a reference, and should certainly not be higher than zero-to-hero.

---

modifications:

1. modified left side nav for devs to promote walkthroughs over
   references, remove incomplete placeholders and bring zero-to-hero to
   the top
2. modified quickstart (first page they see) to promote same
3. promoted “guided walkthroughs” to introduction page and re-ordered
   to put zero-to-hero first
4. modified references section to include contract and project structure
   pages, as per your feedback and one existing page