fix: link sidebar correctly to near-sdk-rs
And add front matter to near-sdk-rs markdown file so that it shows up
with the same "Rust" in the sidebar

I would LOVE a way to override the name of the sidebar item _in
sidebars.json_, but I couldn't find a way to do this. That would let us
match the hierarchy of "AssemblyScript -> README" on the API page and
keep the current "Rust" name on the Develop sidebar.

Also, something is broken here! The link to Rust no longer shows up on
the Develop tab. 😞  I can't figure out why, though.