Full zero to hero update (V0 - need user testing)
I've simplified the tutorial a lot so that it only deals with the
Bitcoin price. We can add a follow up tutorial on how to make this
general purpose, but I think that making this particular one shorter
will be good user experience since there will be more people getting to
the "hero" stage.