fix: create-near-app now defaults to Vanilla
`create-near-app` is being updated to default to vanilla JS in near/create-near-app#387

there have been some other changes lately, too, and this updates the
docs to match

additionally, I doubt it makes sense to have an entire dedicated page
for `create-near-app` in the docs, so I've removed it. This sets us up
to have duplicated docs here and in the `create-near-app` README. This
moves us closer to a single source of truth and makes the docs easier to
maintain.