fix seat price math and tune number of seats wording
Initially spotted in https://near.org/blog/near-protocol-economics/
and then in this document. I think seat price is 21,200.

      $ bc
      1000000 + 500000 + 300000 + 300000 + 20000
      2120000
      48 + 24 + 14 + 14
      100
      2120000 / 100
      21200

Also I found wording confusing, as when one thinks about division,
there is dividend and divisor and sentence was missing explanation
what is that divisor.

With another part of the same sentene I've changed

> you will get required number of seats

to

> you will get assigned number of seats

as in my mind required number of seats is defined by the blockchain
configuration and assigned would be just in that epoch, just for that
validator.