Hotfix sandbox link (#765)
There was a link check on CI but seems no such check now. But there's a link check on master and this two links were broken, because this PR is not yet merged: https://github.com/near/docs/pull/706. So I remove these two links for now and will add back after sandbox doc is merged.