Update documentation to mention `make neard`
We are planning to disable LTO in Cargo.toml which means that
`cargo build --release` will produce less optimised binaries (but
do it faster).  Point at `make neard` in the documentation to make
sure that anyone intending to run node in production gets best
optimised version as before.

The `make neard` target produces the `neard` binary but it has been
added only recently† and is not available when building older
releases.  Because of that, when documentation mentions checking out
non-HEAD commits points at `make release` instead.  It builds a few
more targets so it is slower but it does produce `neard`.

Note that docs/community/contribute/nearcore.md file has been left
unchanged because it describes development process (rather than
running the node in production) and in that case we want people to
build via `cargo build` et al which is much faster‡.

† https://github.com/near/nearcore/pull/4510
‡ https://github.com/near/nearcore/pull/4507

Issue: https://github.com/near/nearcore/issues/4403