Restore data backup to expected directory
When executing the bash commands as documented, all of the files in the backup archive end up extracted into the top level of the near home directory. On first start of neard, it then creates a new data structure in near home/data and begins to sync with peers from the beginning. The updates in this commit extract the archive into the expected /data directory and allow the synchronization to start from the last back up.