truncate sandbox docs, link to near-runner
I was going to update the page about NEAR Sandbox with up-to-date info
about the newer, better way to write tests with
[near-runner](https://github.com/near/runner)

But then it seemed like I was just going to copy-paste its readme
into this page.

I generally think:

* We should avoid copying & pasting most things, definitely readmes
* The further documentation lives from the code it documents, the
  higher-level it should be

And the readme for near-runner is already quite high-level and
approachable!

So I'm not sure what to put on this page anymore. @thisisjoshford what
do you think?