Update versions in ‘Run a Foo Node’ pages
Most importantly, remove the ‘Current latest is 1.20.0’ wording since
it’s out of date.  Updating it would create a burden of maintaining
the version so instead just don’t state what the latest release is and
only point people to the releases page on GitHub.

At the same time, update versions referenced in the documents
to 1.22.0.  This makes it consistent between the pages.

Finally, I’ve also noticed that the estimate of testnet backup is
widely out dated so update that as well.