Improve data snapshot instructions
Firstly, `wget <file> <url>` is an invalid invocation so fix that to
using `-P` option.  (Another alternative would be to use `-O` option
but since existing documents already use `-P` use that consistently in
Node Data Snapshots section as well).  With that, move `-P` at the
front of the command so that it’s more obvious that the switch is
there and is necessary.

Secondly, for better or worse (*cough* worse *cough*) the `data.tar`
files do not include the `data` directory so make sure that
instructions indicate that the file needs to be extracted inside of
the `~/.near/data` directory.