Made cargo new create library instead of main (#1061)
NEAR doesn't use main.rs. In the tutorial it tells us to do a normal `cargo new` and  to delete the main.rs file and add a lib.rs file. `Cargo new` has an option to make it create a library right off the bat. I think it would be less confusing for people to not have to delete and remake a new file right off the bat.