Update providers.md
Hi,

I came across your RPC Providers Resources listing "https://docs.near.org/vi/api/rpc/providers" and found that Zeeve is not listed there. I take this opportunity to bring to your attention that Zeeve has been one of the most popular no code web3 platforms for enterprises and has been providing RPC node hosting services. 

Hereby, I request you to please add Zeeve to the “RPC Provider”  page where all other competitors have also been listed. Zeeve is a cloud-based platform for managing and deploying applications. It provides a comprehensive set of features for developers to quickly and easily deploy their applications to the cloud. Zeeve offers a range of features, including automated deployment of applications to the cloud of your choice, monitoring of applications, automated security and compliance and so on.  

We believe that Zeeve is an excellent addition to the Github directory and would be a great resource for developers looking to quickly and easily deploy their decentralized applications to the cloud.