chore: disable automatic link checker
The link checker is more annoying than anything else since it would randomly fail with some external links. At this point, it has become such a broken feature that we do not even bother to check if the external links are actually broken or not.

On the other hand, we have already activated the `do not build if internal links are broken`, which is way more important (as it means that our documentation keeps internal consistency)

I am for now disabling the automatic run of the link checker. We will see to run it manually once per month, or maybe we can revisit it later to make it more robuts.